# api/sample.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "sample.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/samples": {
      "get": {
        "operationId": "ListSamples",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serviceListSamplesResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "SampleService"
        ]
      },
      "post": {
        "operationId": "CreateSample",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serviceSample"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serviceSample"
            }
          }
        ],
        "tags": [
          "SampleService"
        ]
      }
    },
    "/samples/{sample_id}": {
      "get": {
        "operationId": "GetSample",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serviceSample"
            }
          }
        },
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "SampleService"
        ]
      },
      "delete": {
        "operationId": "DeleteSample",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          }
        },
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "SampleService"
        ]
      },
      "patch": {
        "operationId": "UpdateSample",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serviceSample"
            }
          }
        },
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serviceSample"
            }
          }
        ],
        "tags": [
          "SampleService"
        ]
      }
    }
  },
  "definitions": {
    "serviceListSamplesResponse": {
      "type": "object",
      "properties": {
        "samples": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/serviceSample"
          }
        }
      }
    },
    "serviceSample": {
      "type": "object",
      "properties": {
        "hoge": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "age": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  }
}